requirements.txt
requests>=2.28.2
PyJWT>=2.6.0
Pillow>=9.4.0
numpy>=1.23.5